# chore(lwc): replace terser with esbuild