## Details

Replaces `terser` with `esbuild` for our `*.min.js` files built in the `lwc` package. I'm using the "standalone minify" config described [here](https://github.com/egoist/rollup-plugin-esbuild#standalone-minify-plugin).

This speeds up the `yarn` build from 0m27.716s to 0m23.963s (**~3.7s faster**) (median of 5 runs).

The minified files are slightly larger:

| Filename                                             | Before | After | Delta | Delta (%) |
|------------------------------------------------------|--------|-------|-------|-----------|
| engine-dom/umd/es5/engine-dom.min.js                 | 55738  | 58382 | 2644  | 4.7%      |
| engine-dom/umd/es2017/engine-dom.min.js              | 48370  | 50668 | 2298  | 4.8%      |
| engine-dom/iife/es5/engine-dom.min.js                | 55547  | 58204 | 2657  | 4.8%      |
| engine-dom/iife/es2017/engine-dom.min.js             | 48179  | 50490 | 2311  | 4.8%      |
| synthetic-shadow/umd/es5/synthetic-shadow.min.js     | 46380  | 46913 | 533   | 1.1%      |
| synthetic-shadow/umd/es2017/synthetic-shadow.min.js  | 44284  | 44746 | 462   | 1.0%      |
| synthetic-shadow/iife/es5/synthetic-shadow.min.js    | 46314  | 46847 | 533   | 1.2%      |
| synthetic-shadow/iife/es2017/synthetic-shadow.min.js | 44218  | 44679 | 461   | 1.0%      |
| engine-server/commonjs/es2017/engine-server.min.js   | 45702  | 47817 | 2115  | 4.6%      |
| wire-service/umd/es5/wire-service.min.js             | 4597   | 4601  | 4     | 0.1%      |
| wire-service/umd/es2017/wire-service.min.js          | 2504   | 2522  | 18    | 0.7%      |
| wire-service/iife/es5/wire-service.min.js            | 4406   | 4427  | 21    | 0.5%      |
| wire-service/iife/es2017/wire-service.min.js         | 2313   | 2348  | 35    | 1.5%      |


I'm not really bothered by this, since in core we don't use the `*.min.js` files anyway. And if folks are bothered by the file size increase, then they can minify themselves.

Arguably we should be doing #2430 and remove the minified files entirely, but that would be a breaking change. So instead, I think this is a nice short-term fix to make the build slightly faster.

## Does this pull request introduce a breaking change?

<!--
    Any change that can cause downstream consumers to fail qualifies as a breaking change.
    
    Examples:
        - Removing the code for a deprecated API.
        - Adding a new restriction to the compiler which might result in a compilation failure for existing code.
        - Changing the return type of a function in a non-backward compatible fashion.

    Remove the incorrect item for the list. 
-->
* ✅ No, it does not introduce a breaking change.

<!-- If yes, please describe the impact and migration path for existing applications. -->

## Does this pull request introduce an observable change?

<!--
    Observable changes are internal changes that can be observed by downstream consumers. 
    Such changes don't qualify as breaking changes because they don't impact any publicly defined 
    APIs.

    Examples:
        - Fixing a bug.
        - Changing the invocation timing of a callback, for a callback that has no invocation timing
          guarantee.

    Remove the incorrect item from the list. 
-->
* ✅ No, it does not introduce an observable change.

<!-- If yes, please describe the anticipated observable changes. -->